Treat Inotify IN_MOVED_TO events as creations
Both the Default and KQueue watchers treat files being renamed into (or
within) a watched directory as a creation event. However, the Inotify
watcher was ignoring those events. Fix that.

Given a watcher on directory d/ with filter qr/\.txt\z/, this affects
the following operations:

  * Renaming other/file.txt to d/file.txt
  * Renaming d/file.txt-tmp to d/file.txt

The latter of those is particularly useful when a file is created with
a temporary name, then renamed into place. For example, that's what the
"spew" method on Path::Tiny instances does.

This change has been tested on both Mac OS (with KQueue) and Linux (with
Inotify).